requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.3
django-clearcache==1.2.1
django-crispy-forms==1.12.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
gunicorn==20.1.0
idna==2.10
pep8==1.7.1
psycopg2==2.8.6
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.23
whitenoise==5.2.0
